+++ libgdx/gdx/src/com/badlogic/gdx/utils/compression/lzma/Base.java
@@ -62,6 +62,6 @@ public class Base {
     public static final int GetLenToPosState(int len) {
         len -= kMatchMinLen;
         if (len < kNumLenToPosStates) return len;
-        return (int) (kNumLenToPosStates - 1);
+        return kNumLenToPosStates - 1;
     }
 }
